requirements.txt
lxml
jinja2
opencc-python-reimplemented
Pillow
readability-lxml
argostranslate
stanza
tzdata
